// src/app/components/Models/SessionDetails.ts
export class SessionDetails {
    userName: string;
    grant_type: string;
    LstClient: string;
    LstCountry: string;
    LstFunctionalAreas: string;
    LstGenericCode: string;
    LstMandateStatus: string;
    LstMandateTypes: string;
    LstQualification: string;
    LstTargetIndustries: string;
    LstMandateStage: string;
    LocalSessionDetails: string;
    AccessToken: string;
    TokenType: string;
    TokenExpiresIn: string;
    LstUser: string;
    MandatesDetials: string;
    Api_endpoints: string;
    GridLayoutEnable: string;
    LstCityBusinessPartners: string;
    LstBusinessPartners: string;
    LstBranch: string;
}